added makefile + multiple src files
