+++ src/tokenizer.h
@@ -1,0 +1,8 @@
+#ifndef TOEKNIZER_H_INCLUDED
+#define TOEKNIZER_H_INCLUDED
+
+using namespace std;
+
+void printtest(string input);
+
+#endif // TOEKNIZER_H_INCLUDED
