+++ zinc_testing.cpp
